BS-9304 : [6.3] Temporary folder never get cleared
- Remove some temporary files for tests

+++ bpm/bonita-api/bonita-common-api/src/main/java/org/bonitasoft/engine/io/IOUtil.java
@@ -23,7 +23,6 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStreamWriter;
-import java.io.Writer;
 import java.net.URI;
 import java.net.URL;
 import java.nio.MappedByteBuffer;
@@ -49,7 +48,7 @@ public class IOUtil {
 
     private static final String LINE_SEPARATOR = System.getProperty("line.separator");
 
-    private static final String TMP_DIRECTORY = System.getProperty("java.io.tmpdir");
+    public static final String TMP_DIRECTORY = System.getProperty("java.io.tmpdir");
 
     private static final int BUFFER_SIZE = 100000;
 
@@ -243,23 +242,28 @@ public class IOUtil {
     }
 
     public static boolean deleteDir(final File dir, final int attempts, final long sleepTime) throws IOException {
-        boolean result = true;
-        if (!dir.exists()) {
-            return false;
-        }
-        if (!dir.isDirectory()) {
-            throw new IOException("Unable to delete directory: " + dir + ", it is not a directory");
-        }
-        final File[] files = dir.listFiles();
-        for (int i = 0; i < files.length; i++) {
-            if (files[i].isDirectory()) {
-                deleteDir(files[i], attempts, sleepTime);
-            } else {
-                result = result && deleteFile(files[i], attempts, sleepTime);
+        if (dir != null) {
+            boolean result = true;
+            if (!dir.exists()) {
+                return false;
             }
+            if (!dir.isDirectory()) {
+                throw new IOException("Unable to delete directory: " + dir + ", it is not a directory");
+            }
+            for (final File file : dir.listFiles()) {
+                if (file.isDirectory()) {
+                    deleteDir(file, attempts, sleepTime);
+                } else {
+                    result = result && deleteFile(file, attempts, sleepTime);
+                }
+            }
+            return result && deleteFile(dir, attempts, sleepTime);
         }
-        result = result && deleteFile(dir, attempts, sleepTime);
-        return result;
+        return false;
+    }
+
+    public static File createTempFileInDefaultTempDirectory(final String prefix, final String suffix) throws IOException {
+        return createTempFile(prefix, suffix, new File(TMP_DIRECTORY));
     }
 
     public static File createTempFile(final String prefix, final String suffix, final File directory) throws IOException {
@@ -361,28 +365,40 @@ public class IOUtil {
      */
     public static void zipDir(final String dir2zip, final ZipOutputStream zos, final String root) throws IOException {
         final File zipDir = new File(dir2zip);
-        String[] dirList;
-        dirList = zipDir.list();
         final byte[] readBuffer = new byte[BUFFER_SIZE];
         int bytesIn = 0;
-        for (int i = 0; i < dirList.length; i++) {
-            final File f = new File(zipDir, dirList[i]);
-            final String path = f.getPath();
-            if (f.isDirectory()) {
-                final String filePath = path;
-                zipDir(filePath, zos, root);
+
+        for (final String pathName : zipDir.list()) {
+            final File file = new File(zipDir, pathName);
+            final String path = file.getPath();
+            if (file.isDirectory()) {
+                zipDir(path, zos, root);
                 continue;
             }
-            final FileInputStream fis = new FileInputStream(f);
-            final ZipEntry anEntry = new ZipEntry(path.substring(root.length() + 1, path.length()).replace(String.valueOf(File.separatorChar), "/"));
-            zos.putNextEntry(anEntry);
+
+            try {
+                final ZipEntry anEntry = new ZipEntry(path.substring(root.length() + 1, path.length()).replace(String.valueOf(File.separatorChar), "/"));
+                zos.putNextEntry(anEntry);
+                bytesIn = +copyFileToZip(zos, readBuffer, file, bytesIn);
+                zos.flush();
+            } finally {
+                zos.closeEntry();
+            }
+        }
+    }
+
+    private static int copyFileToZip(final ZipOutputStream zos, final byte[] readBuffer, final File file, final int bytesInOfZip) throws FileNotFoundException,
+            IOException {
+        final FileInputStream fis = new FileInputStream(file);
+        int bytesIn = bytesInOfZip;
+        try {
             while ((bytesIn = fis.read(readBuffer)) != -1) {
                 zos.write(readBuffer, 0, bytesIn);
             }
+        } finally {
             fis.close();
-            zos.flush();
-            zos.closeEntry();
         }
+        return bytesIn;
     }
 
     /**
@@ -428,29 +444,21 @@ public class IOUtil {
      */
     public static String read(final File file) throws IOException {
         FileInputStream fileInputStream = null;
-        Scanner scanner = null;
         try {
             fileInputStream = new FileInputStream(file);
-            scanner = new Scanner(fileInputStream, FILE_ENCODING);
-            return read(scanner);
+            return read(fileInputStream);
         } finally {
             if (fileInputStream != null) {
                 fileInputStream.close();
-            }
-            if (scanner != null) {
-                scanner.close();
             }
         }
     }
 
     public static void unzipToFolder(final InputStream inputStream, final File outputFolder) throws IOException {
         final ZipInputStream zipInputstream = new ZipInputStream(inputStream);
-        ZipEntry zipEntry = null;
 
         try {
-            while ((zipEntry = zipInputstream.getNextEntry()) != null) {
-                extractZipEntry(zipInputstream, zipEntry, outputFolder);
-            }
+            extractZipEntries(zipInputstream, outputFolder);
         } finally {
             zipInputstream.close();
         }
@@ -463,24 +471,23 @@ public class IOUtil {
         return true;
     }
 
-    private static void extractZipEntry(final ZipInputStream zipInputstream, final ZipEntry zipEntry, final File outputFolder) throws FileNotFoundException,
+    private static void extractZipEntries(final ZipInputStream zipInputstream, final File outputFolder) throws FileNotFoundException,
             IOException {
-        try {
-            final String entryName = zipEntry.getName();
-            // entryName = entryName.replace('/', File.separatorChar);
-            // entryName = entryName.replace('\\', File.separatorChar);
+        ZipEntry zipEntry = null;
+        while ((zipEntry = zipInputstream.getNextEntry()) != null) {
+            try {
+                // For each entry, a file is created in the output directory "folder"
+                final File outputFile = new File(outputFolder.getAbsolutePath(), zipEntry.getName());
 
-            // For each entry, a file is created in the output directory "folder"
-            final File outputFile = new File(outputFolder.getAbsolutePath(), entryName);
-
-            // If the entry is a directory, it creates in the output folder, and we go to the next entry (return).
-            if (zipEntry.isDirectory()) {
-                mkdirs(outputFile);
-                return;
+                // If the entry is a directory, it creates in the output folder, and we go to the next entry (return).
+                if (zipEntry.isDirectory()) {
+                    mkdirs(outputFile);
+                    return;
+                }
+                writeZipInputToFile(zipInputstream, outputFile);
+            } finally {
+                zipInputstream.closeEntry();
             }
-            writeZipInputToFile(zipInputstream, outputFile);
-        } finally {
-            zipInputstream.closeEntry();
         }
     }
 
@@ -496,6 +503,7 @@ public class IOUtil {
                 while ((bytesRead = zipInputstream.read(buffer)) > -1) {
                     fileOutputStream.write(buffer, 0, bytesRead);
                 }
+                fileOutputStream.flush();
             } finally {
                 fileOutputStream.close();
             }
@@ -507,11 +515,19 @@ public class IOUtil {
     }
 
     public static void writeContentToFile(final String content, final File outputFile) throws IOException {
-        final Writer out = new OutputStreamWriter(new FileOutputStream(outputFile), FILE_ENCODING);
+        final FileOutputStream fileOutput = new FileOutputStream(outputFile);
+        OutputStreamWriter out = null;
         try {
+            out = new OutputStreamWriter(fileOutput, FILE_ENCODING);
             out.write(content);
+            out.flush();
         } finally {
-            out.close();
+            if (out != null) {
+                out.close();
+            }
+            if (fileOutput != null) {
+                fileOutput.close();
+            }
         }
     }
 
@@ -522,6 +538,7 @@ public class IOUtil {
             fos = new FileOutputStream(file);
             bos = new BufferedOutputStream(fos);
             bos.write(entry.getValue());
+            bos.flush();
         } finally {
             if (bos != null) {
                 bos.close();
@@ -544,11 +561,11 @@ public class IOUtil {
             buf.get(bytes);
             return bytes;
         } finally {
-            if (fin != null) {
-                fin.close();
-            }
             if (ch != null) {
                 ch.close();
+            }
+            if (fin != null) {
+                fin.close();
             }
         }
     }
